requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
tenacity==8.2.3
httpx>=0.24.0 
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.2.1
google-cloud-bigquery==3.19.0
schedule==1.2.1
textstat==0.7.3
chromadb==0.4.24
sentence-transformers==2.5.1
langchain==0.1.12 